requirements.txt
customtkinter 
pandas
brapi 
matplotlib.pyplot 
sqlite3